+++ src/interfaces/ISkillOptions.ts
@@ -1,4 +1,0 @@
-export default interface ISKillOptions {
-  event: any;
-  additionalMods?: (string | number)[];
-}